+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-worker
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Use simplejson from django.utils.
 
 
 0.1 (2012-10-16)